consolidate server actions code
